+++ src/components/Icon/icons/IconCopy/index.js
@@ -3,15 +3,15 @@ import * as React from 'react';
 import { css, cx } from 'emotion';
 import { Icon, type GenericIconProps } from '../../Icon';
 import { colors } from '../../../../variables';
-import image from './new-window.svg';
+import image from './icon-copy.svg';
 
 const styles = css`
-  width: 12px;
-  height: 12px;
-  fill: ${colors.intentPrimary65};
+  width: 16px;
+  height: 16px;
+  fill: ${colors.dark65};
 `;
 
-export const IconNewWindow = ({ className }: GenericIconProps) => (
+export const IconCopy = ({ className }: GenericIconProps) => (
   <Icon
     className={cx(styles, className)}
     glyph={image}